# Backend/documentation/store/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/products/{product_id}/logo": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update Product Logo",
        "operationId": "updateProductLogo",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439042"
            },
            "example": "507f1f77bcf86cd799439042"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "logo"
                ],
                "type": "object",
                "properties": {
                  "logo": {
                    "type": "string",
                    "description": "The logo image file (PNG, JPEG, SVG or WEBP) - max 10MB",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product logo was updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoUpdatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "File format is incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          },
          "413": {
            "description": "File is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get Product List",
        "operationId": "getProductList",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439044"
            },
            "example": "507f1f77bcf86cd799439044"
          }
        ],
        "responses": {
          "200": {
            "description": "Store is found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductShortInfo"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Store is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create Product",
        "operationId": "createProduct",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "store_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439044"
            },
            "example": "507f1f77bcf86cd799439044"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Store is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/products/{product_id}": {
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "Update Product",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439042"
            },
            "example": "507f1f77bcf86cd799439042"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/products/{product_id}/short": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get Product Short Info",
        "operationId": "getProductInfoShort",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439042"
            },
            "example": "507f1f77bcf86cd799439042"
          }
        ],
        "responses": {
          "200": {
            "description": "Product is found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductShortInfo"
                }
              }
            }
          },
          "404": {
            "description": "Product is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/products/{product_id}/full": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get Product Full Info",
        "operationId": "getProductInfoFull",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439042"
            },
            "example": "507f1f77bcf86cd799439042"
          }
        ],
        "responses": {
          "200": {
            "description": "Product is found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullInfo"
                }
              }
            }
          },
          "404": {
            "description": "Product is not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LogoUpdatedResponse": {
        "required": [
          "logo_url"
        ],
        "type": "object",
        "properties": {
          "logo_url": {
            "type": "string",
            "description": "Logo image",
            "format": "uri",
            "example": "https://i.pinimg.com/736x/5a/38/a6/5a38a6077eb3115b5ab1da7825d80032.jpg"
          }
        }
      },
      "DetailResponse": {
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "example": "detailed explanation of result"
          }
        }
      },
...
      "ProductCreatedResponse": {
        "required": [
          "product_id"
        ],
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "description": "Id of the product",
            "example": "507f1f77bcf86cd799439042"
          }
        }
      },
      "Characteristic": {
        "required": [
          "name",
          "value"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the characteristic",
            "example": "Characteristic name"
          },
          "value": {
            "type": "string",
            "description": "Value of the characteristic",
            "example": "Characteristic nam value"
          }
        },
        "description": "Characteristics of the product"
      },
      "CreateProductModel": {
        "required": [
          "characteristics",
          "description",
          "price",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the product",
            "example": "Product title"
          },
          "description": {
            "type": "string",
            "description": "Description of the product",
            "example": "Product description"
          },
          "price": {
            "type": "integer",
            "description": "Price of the product in kopecks",
            "format": "int32",
            "example": 100000
          },
          "characteristics": {
            "type": "array",
            "description": "Characteristics of the product",
            "items": {
              "$ref": "#/components/schemas/Characteristic"
            }
          }
        }
      },
...
      "UpdateProductModel": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the product",
            "example": "Product title"
          },
          "description": {
            "type": "string",
            "description": "Description of the product",
            "example": "Product description"
          },
          "price": {
            "type": "integer",
            "description": "Price of the product in kopecks",
            "format": "int32",
            "example": 100000
          },
          "characteristics": {
            "type": "array",
            "description": "Characteristics of the product",
            "items": {
              "$ref": "#/components/schemas/Characteristic"
            }
          }
        }
      },
...
      "ProductShortInfo": {
        "required": [
          "price",
          "product_id",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the product",
            "example": "Product title"
          },
          "price": {
            "type": "integer",
            "description": "Price of the product in kopecks",
            "format": "int32",
            "example": 100000
          },
          "product_id": {
            "type": "string",
            "description": "Id of the product",
            "example": "507f1f77bcf86cd799439042"
          },
          "logo_url": {
            "type": "string",
            "description": "Image of the product",
            "format": "uri",
            "example": "https://i.pinimg.com/736x/5a/38/a6/5a38a6077eb3115b5ab1da7825d80032.jpg"
          }
        }
      },
      "ProductFullInfo": {
        "required": [
          "characteristics",
          "description",
          "price",
          "product_id",
          "store_id",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the product",
            "example": "Product title"
          },
          "description": {
            "type": "string",
            "description": "Description of the product",
            "example": "Product description"
          },
          "price": {
            "type": "integer",
            "description": "Price of the product in kopecks",
            "format": "int32",
            "example": 100000
          },
          "characteristics": {
            "type": "array",
            "description": "Characteristics of the product",
            "items": {
              "$ref": "#/components/schemas/Characteristic"
            }
          },
          "product_id": {
            "type": "string",
            "description": "Id of the product",
            "example": "507f1f77bcf86cd799439042"
          },
          "logo_url": {
            "type": "string",
            "description": "Image of the product",
            "format": "uri",
            "example": "https://i.pinimg.com/736x/5a/38/a6/5a38a6077eb3115b5ab1da7825d80032.jpg"
          },
          "store_id": {
            "type": "string",
            "description": "Id of the store in which product is sold",
            "example": "507f1f77bcf86cd799439044"
          }
        }
      }
    }
  }
}